Add splitter to 3D
