+++ lib/ReactViews/Map/TerriaViewerWrapper.jsx
@@ -5,9 +5,9 @@ import PropTypes from 'prop-types';
 import Cartesian2 from 'terriajs-cesium/Source/Core/Cartesian2';
 import Styles from './terria-viewer-wrapper.scss';
 
+import CesiumSplitter from './CesiumSplitter';
 import ObserveModelMixin from '../ObserveModelMixin';
 import TerriaViewer from '../../ViewModels/TerriaViewer';
-import ViewerMode from '../../Models/ViewerMode';
 
 const TerriaViewerWrapper = createReactClass({
     displayName: 'TerriaViewerWrapper',
@@ -19,11 +19,6 @@ const TerriaViewerWrapper = createReactClass({
     propTypes: {
         terria: PropTypes.object.isRequired,
         viewState: PropTypes.object.isRequired
-    },
-
-    getInitialState: function() {
-        const scene = this.props.terria.currentViewer.scene;
-        return {value: scene ? scene.imagerySplitPosition : 0.5};
     },
 
     componentDidMount() {
@@ -61,25 +56,14 @@ const TerriaViewerWrapper = createReactClass({
         }
     },
 
-    onSliderMove(event) {
-        this.setState({value: event.target.value});
-        this.props.terria.currentViewer.scene.imagerySplitPosition = event.target.value;
-    },
-
     render() {
-        const terria = this.props.terria;
         return (
             <aside id="cesiumContainer"
                    className={Styles.cesiumContainer}
                    ref={element => {this.mapElement = element;}}
                    onMouseMove={this.onMouseMove}>
                 <div className={Styles.mapPlaceholder}>Loading the map, please wait!</div>
-                <If condition={terria.viewerMode === ViewerMode.CesiumTerrain || terria.viewerMode === ViewerMode.CesiumEllipsoid}>
-                    <div className="cesiumSplitter">
-                        <div className="leaflet-sbs-divider" style={{left: this.state.value*100 + "%"}}></div>
-                        <input className="leaflet-sbs-range" type="range" min="0" max="1" step="any" value={this.state.value} onChange={this.onSliderMove}/>
-                    </div>
-                </If>
+                <CesiumSplitter terria={this.props.terria} />
             </aside>
         );
     },
