+++ lib/Sass/common/_variables.scss
@@ -146,8 +146,8 @@ $chart-line-color: #ccc;
 $chart-preview-line-color: $chart-line-color;
 $chart-darker: darken($dark, 8%);
 
-
-
 $charcoal-grey: #3f4854;
 $turquoise-blue: #08abd5;
 
+// Splitter
+$splitter-thumb-width: 40px;
