+++ lib/Sass/global/global.scss
@@ -1,2 +1,3 @@
 @import './leaflet';
+@import './splitter';
 @import './normalize';
